requirements.txt
requests==2.31.0
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.40
schedule==1.2.0
python-dotenv==1.0.0

